requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
curlify==3.0.0
distro==1.9.0
Django==6.0
django-cors-headers==4.9.0
django-filter==25.2
django-silk==5.4.3
django-stubs==5.2.8
django-stubs-ext==5.2.8
djangorestframework==3.16.1
djangorestframework-stubs==3.16.6
djangorestframework_simplejwt==5.5.1
facebook_business==24.0.1
frozenlist==1.8.0
gprof2dot==2025.4.14
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
librt==0.7.7
multidict==6.7.0
mypy==1.19.1
mypy_extensions==1.1.0
openai==2.15.0
pathspec==0.12.1
pillow==12.0.0
propcache==0.4.1
pycountry==24.6.1
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.4
tiktok-business-api-sdk-official==1.1.1
tqdm==4.67.1
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.0
yarl==1.22.0
